Improve UT and fix code bug
