+++ common/decoratortable.h
@@ -13,19 +13,20 @@
 
 namespace swss {
 
-class OverlayTable : public Table  
+class DecoratorTable : public Table  
 {
 public:
-    OverlayTable(const DBConnector *db, const std::string &tableName);
-    OverlayTable(RedisPipeline *pipeline, const std::string &tableName, bool buffered);
-    ~OverlayTable() override;
+    DecoratorTable(const DBConnector *db, const std::string &tableName);
+    DecoratorTable(RedisPipeline *pipeline, const std::string &tableName, bool buffered);
+    ~DecoratorTable() override;
 
     /* Get all the field-value tuple of the table entry with the key */
     bool get(const std::string &key, std::vector<std::pair<std::string, std::string>> &ovalues) override;
 
+    /* Get an entry field-value from the table */
     bool hget(const std::string &key, const std::string &field,  std::string &value) override;
 
-    /* get all the keys in the table */
+    /* Get all the keys from the table */
     void getKeys(std::vector<std::string> &keys) override;
 
     /* Set an entry in the DB directly and configure ttl for it (op not in use) */
@@ -40,6 +41,7 @@ public:
                      const std::string &op = "",
                      const std::string &prefix = EMPTY_PREFIX) override;
 
+    /* Set an entry field in the table */
     void hset(const std::string &key,
                      const std::string &field,
                      const std::string &value,
